requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
phonenumberslite==9.0.16
PyMySQL==1.1.2
Werkzeug==3.1.3
cryptography>=42.0.0 
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0